+++ src/modules/innovation_mod.f90
@@ -483,7 +483,7 @@ contains
     ! Locals:
     real(8) :: zjo,zjoraob,zjosatwind,zjosurfc
     real(8) :: zjosfcsf,zjosfcua,zjotov,zjoairep,zjosfcsc,zjoprof,zjoaladin,zjosfctm
-    real(8) :: zjogpsro,zjogpsgb,zjosfcgp,zjosfcra,zjochm,zjosfcgl,zjosfchy
+    real(8) :: zjogpsro,zjogpsgb,zjosfcgp,zjosfcra,zjochm,zjosfcgl,zjosfchy,zjoradvel
     integer :: destObsColumn, get_max_rss
     logical :: lgpdata, beSilent
 
@@ -547,6 +547,10 @@ contains
     call oop_sfc_nl  (columnhr, obsSpaceData, beSilent, ZJOSFCSC, 'SC', destObsColumn)
     call oop_sfc_nl  (columnhr, obsSpaceData, beSilent, ZJOSFCGP, 'GP', destObsColumn)
     call oop_sfc_nl  (columnhr, obsSpaceData, beSilent, ZJOSFCRA, 'RA', destObsColumn)
+    !
+    !        RADAR Doppler Velocity
+    !--------------------------------
+    call oop_raDvel_nl (columnhr,obsSpaceData, beSilent,ZJORADVEL,'RA', destObsColumn)  
     !
     !        SEA SURFACE TEMPERATURE
     !--------------------------------
@@ -633,12 +637,14 @@ contains
       write(*,'(a15,f30.16)') 'JOGPSRO  = ',ZJOGPSRO
       write(*,'(a15,f30.16)') 'JOGPSGB  = ',ZJOGPSGB
       write(*,'(a15,f30.16)') 'JOCHM    = ',ZJOCHM
+      write(*,'(a15,f30.16)') 'JORADVEL = ',ZJORADVEL
     end if
 
     !
     !=======================================================================
     ZJO =  ZJORAOB + ZJOAIREP + ZJOSATWIND + &
-         ZJOSURFC + ZJOTOV + ZJOPROF + ZJOALADIN + ZJOGPSRO + ZJOGPSGB + ZJOCHM
+         ZJOSURFC + ZJOTOV + ZJOPROF + ZJOALADIN + ZJOGPSRO + ZJOGPSGB + ZJOCHM + ZJORADVEL
+
     !=======================================================================
 
     if ( .not.beSilent ) then
@@ -661,6 +667,7 @@ contains
       call mpi_allreduce_sumreal8scalar(ZJOCHM,'GRID')
       call mpi_allreduce_sumreal8scalar(ZJOSFCTM,'GRID')
       call mpi_allreduce_sumreal8scalar(ZJOSFCGL,'GRID')
+      call mpi_allreduce_sumreal8scalar(ZJORADVEL,'GRID')
 
       write(*,*) 'Cost function values summed for all MPI tasks:'
       write(*,'(a15,f30.16)') 'JORAOB   = ',ZJORAOB
@@ -680,7 +687,7 @@ contains
       write(*,'(a15,f30.16)') 'JOGPSRO  = ',ZJOGPSRO
       write(*,'(a15,f30.16)') 'JOGPSGB  = ',ZJOGPSGB
       write(*,'(a15,f30.16)') 'JOCHM    = ',ZJOCHM
-
+      write(*,'(a15,f30.16)') 'JORADAR  = ',ZJORADVEL
     end if ! beSilent
 
     call mpi_allreduce_sumreal8scalar(ZJO,'GRID')
@@ -820,7 +827,7 @@ contains
     ! Locals:
     integer :: numPerturbations
     integer :: nrandseed,iseed,indexAnalysis2,indexBody,indexFamily
-    integer, parameter :: numFamily=9
+    integer, parameter :: numFamily=10
     real*8  :: zmean,originalOmp
     real*8  :: scaleFactor(numFamily)
     character(len=2) :: familyList(numFamily)
@@ -836,6 +843,7 @@ contains
     familyList(7)='PR' ; scaleFactor(7)=1.00d0
     familyList(8)='RO' ; scaleFactor(8)=1.00d0
     familyList(9)='GP' ; scaleFactor(9)=1.00d0
+    familyList(10)='RA'; scaleFactor(10)=1.00d0    
 
     numPerturbations = numAnalyses